requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.6
django-cors-headers==3.8.0
django-environ-2==2.1.0
django-ipware==3.0.7
djangorestframework==3.12.4
drf-recaptcha==2.0.4
google-api-core==2.0.1
google-auth==2.2.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.1
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
idna==3.2
mysqlclient==2.0.3
numpy==1.21.4
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
